+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to the "erabasic" extension will be documented in this file.
 
 ## [Unreleased]
 ### Added
+- Simple Code Completion: L10N に対応
 - Go to Definition/Symbols: UTF-8, UTF-16 の BOM 判定を追加
 
 ## [0.3.0] - 2020-05-12